InstructorsGuide/ai-agents-hol-setup/3_neo4j_setup/neo4j_kaggle_brazilian_ecommerce/requirements.txt
neo4j
pandas
python-dotenv
